+++ cookbooks/app-react-bookmark-advanced/src/Routes.tsx
@@ -29,7 +29,7 @@ export const routes: RouteObject[] = [
                 index: true,
                 element: (
                     <div>
-                        <Selected />,
+                        <Selected />
                     </div>
                 ),
             },