+++ cookbooks/app-react-module/src/modules/demo/index.ts
@@ -1,0 +1,2 @@
+export { type DemoModule, demoModule } from './module';
+export { type DemoModuleConfig } from './configurator';