+++ packages/modules/context/src/utils/index.ts
@@ -1,0 +1,2 @@
+export { enableContext } from './enable-context';
+export { resolveInitialContext } from './resolve-initial-context';