+++ packages/modules/services/src/provider.ts
@@ -33,6 +33,44 @@ type ApiProviderCtorArgs<TClient extends IHttpClient = IHttpClient> = {
     createClient: ApiClientFactory<TClient>;
 };
 
+// TODO move to own file
+type ApiProviderErrorResponse = {
+    type: ResponseType;
+    status: number;
+    statusText: string;
+    headers: Headers;
+    url: string;
+    data: unknown;
+};
+
+// TODO move to own file
+export class ApiProviderError extends Error {
+    readonly response: ApiProviderErrorResponse;
+
+    constructor(msg: string, response: ApiProviderErrorResponse, options?: ErrorOptions) {
+        super(msg, options);
+        this.response = response;
+        this.name = 'ApiProviderError';
+    }
+}
+
+// TODO move to own file
+const validateResponse = async (response: Response) => {
+    if (!response.ok) {
+        const { headers, status, statusText, type, url, bodyUsed } = response;
+        const isJson = headers.get('content-type')?.match(/application\/(\w*)?[+]?json/);
+        const data = !bodyUsed && (await (isJson ? response.json() : response.text()));
+        throw new ApiProviderError('failed to execute request, response was not ok', {
+            headers,
+            status,
+            statusText,
+            type,
+            url,
+            data,
+        });
+    }
+};
+
 export class ApiProvider<TClient extends IHttpClient = IHttpClient>
     implements IApiProvider<TClient>
 {
@@ -45,6 +83,7 @@ export class ApiProvider<TClient extends IHttpClient = IHttpClient>
         method: TMethod
     ): Promise<NotificationApiClient<TMethod, TClient>> {
         const httpClient = await this._createClientFn('notification');
+        httpClient.responseHandler.add('validate_api_request', validateResponse);
         return new NotificationApiClient(httpClient, method);
     }
 
@@ -52,6 +91,7 @@ export class ApiProvider<TClient extends IHttpClient = IHttpClient>
         method: TMethod
     ): Promise<BookmarksApiClient<TMethod, TClient, TPayload>> {
         const httpClient = await this._createClientFn('bookmarks');
+        httpClient.responseHandler.add('validate_api_request', validateResponse);
         return new BookmarksApiClient(httpClient, method);
     }
 
@@ -59,6 +99,7 @@ export class ApiProvider<TClient extends IHttpClient = IHttpClient>
         method: TMethod
     ): Promise<ContextApiClient<TMethod, TClient>> {
         const httpClient = await this._createClientFn('context');
+        httpClient.responseHandler.add('validate_api_request', validateResponse);
         return new ContextApiClient(httpClient, method);
     }
 }