+++ packages/react/modules/bookmark/src/portal/useBookmarkNavigate.ts
@@ -40,8 +40,8 @@ export const useBookmarkNavigate = (args: { resolveAppPath: AppPathResolver }): 
             }
 
             if (bookmarkContext) {
-                context.contextClient.currentContext?.id !== bookmarkContext.id &&
-                    context.contextClient.setCurrentContext(bookmarkContext.id);
+                context.currentContext?.id !== bookmarkContext.id &&
+                    context.setCurrentContextByIdAsync(bookmarkContext.id);
             }
         });
         return sub;