+++ packages/modules/telemetry/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 3.0.6 (2023-05-23)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-telemetry
+
+## 3.0.5 (2023-05-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-telemetry
+
 ## 3.0.4 (2023-04-18)
 
 ### Bug Fixes